Replace list of keyword arguments with dictionary
Made this change for my own purposes and thought I should share it here in case others prefer to use a dictionary rather than manually name all the variables you want to export. Haven't comprehensively tested it or anything but it worked for me. This changes the order of arguments in datax() to avoid positional/keyword arg errors, such that now the first argument is the dictionary of variables, and the second argument is the filename string.